refactor: split dashboard into modular components for better performance
- Create reusable dashboard components directory
- Add Sidebar component with user profile, navigation, and theme toggle
- Add TranscriptionDisplay component for text display and editing
- Add ControlPanel component for mode switching and controls
- Add RecordingControls component for audio recording functionality
- Add UploadInterface component for file upload with drag & drop
- Refactor main dashboard page to use new components
- Improve code maintainability and readability
- Reduce bundle size and enable better code splitting
- All components are TypeScript-typed and SSR-safe
