+++ Frontend/src/app/dashboard/page.tsx
@@ -3,43 +3,12 @@
 import { useState, useRef, useEffect, useCallback } from 'react';
 import { createClient } from '../../lib/supabase/client';
 import { useRouter } from 'next/navigation';
-import { useTheme } from 'next-themes';
-import { Button } from '../../components/ui/button';
-import Image from 'next/image';
-import { 
-  Mic, 
-  Upload, 
-  Download, 
-  Play, 
-  Pause, 
-  Square, 
-  FileAudio, 
-  History, 
-  Settings, 
-  LogOut,
-  Home,
-  Trash2,
-  Copy,
-  Check,
-  Plus,
-  X,
-  Sun,
-  Moon,
-  Menu,
-  MessageCircle,
-  Edit3,
-  Save,
-  RotateCcw
-} from 'lucide-react';
+import { User } from '@supabase/supabase-js';
+import Sidebar from '@/components/dashboard/Sidebar';
+import TranscriptionDisplay from '@/components/dashboard/TranscriptionDisplay';
+import ControlPanel from '@/components/dashboard/ControlPanel';
 
-interface User {
-  id: string;
-  email?: string;
-  user_metadata?: {
-    full_name?: string;
-    avatar_url?: string;
-  };
-}
+type ActiveMode = 'record' | 'chat' | 'upload';
 
 export default function Dashboard() {
   const [user, setUser] = useState<User | null>(null);
@@ -50,17 +19,15 @@ export default function Dashboard() {
   const [isEditing, setIsEditing] = useState(false);
   const [editableText, setEditableText] = useState('');
   const [isCopied, setIsCopied] = useState(false);
-  const [activeMode, setActiveMode] = useState<'record' | 'chat' | 'upload'>('record');
+  const [activeMode, setActiveMode] = useState<ActiveMode>('record');
   const [sidebarOpen, setSidebarOpen] = useState(false);
   const [mounted, setMounted] = useState(false);
   const [isProcessing, setIsProcessing] = useState(false);
   const [audioBlob, setAudioBlob] = useState<Blob | null>(null);
-  const [audioChunks, setAudioChunks] = useState<Blob[]>([]);
   const [uploadedFileName, setUploadedFileName] = useState<string>('');
   const [isDragOver, setIsDragOver] = useState(false);
   const [apiStatus, setApiStatus] = useState<'idle' | 'connecting' | 'connected' | 'error'>('idle');
-  const { resolvedTheme, setTheme } = useTheme();
-  
+
   const supabase = createClient();
   const router = useRouter();
   const mediaRecorderRef = useRef<MediaRecorder | null>(null);
@@ -93,7 +60,7 @@ export default function Dashboard() {
         } else {
           setApiStatus('error');
         }
-      } catch (error) {
+      } catch (err) {
         console.log('API not available, using mock mode');
         setApiStatus('error');
       }
@@ -355,13 +322,11 @@ export default function Dashboard() {
       
       // Reset audio chunks for new recording
       const chunks: Blob[] = [];
-      setAudioChunks(chunks);
       
       // Handle data available event
       mediaRecorder.ondataavailable = (event) => {
         if (event.data.size > 0) {
           chunks.push(event.data);
-          setAudioChunks([...chunks]);
         }
       };
       
@@ -437,12 +402,6 @@ export default function Dashboard() {
     router.push('/');
   };
 
-  const menuItems = [
-    { id: 'home', label: 'Home', icon: Home },
-    { id: 'history', label: 'History', icon: History },
-    { id: 'settings', label: 'Settings', icon: Settings },
-  ];
-
   if (!user) {
     return (
       <div className="min-h-screen bg-gray-50 dark:bg-[#0e0f16] flex items-center justify-center">
@@ -453,409 +412,58 @@ export default function Dashboard() {
 
   return (
     <div className="min-h-screen bg-gray-50 dark:bg-[#0e0f16] flex">
-      {/* Mobile Menu Button */}
-      <div className="lg:hidden fixed top-4 left-4 z-50">
-        <Button
-          onClick={() => setSidebarOpen(!sidebarOpen)}
-          variant="outline"
-          size="sm"
-          className="bg-white dark:bg-[#0a0b0f] border-gray-300 dark:border-white/10 text-gray-900 dark:text-white hover:bg-gray-100 dark:hover:bg-white/10"
-        >
-          <Menu className="w-4 h-4" />
-        </Button>
-      </div>
-
-      {/* Mobile Overlay */}
-      {sidebarOpen && (
-        <div 
-          className="lg:hidden fixed inset-0 bg-black/50 z-40"
-          onClick={() => setSidebarOpen(false)}
-        />
-      )}
-
-      {/* Left Sidebar */}
-      <div className={`
-        w-64 bg-white dark:bg-[#0a0b0f] border-r border-gray-200 dark:border-white/10 flex flex-col
-        lg:relative lg:translate-x-0
-        fixed inset-y-0 left-0 z-50 transform transition-transform duration-300 ease-in-out
-        ${sidebarOpen ? 'translate-x-0' : '-translate-x-full lg:translate-x-0'}
-      `}>
-        {/* Close Button for Mobile */}
-        <div className="lg:hidden p-4 flex justify-end">
-          <Button
-            onClick={() => setSidebarOpen(false)}
-            variant="outline"
-            size="sm"
-            className="bg-gray-100 dark:bg-white/5 border-gray-300 dark:border-white/10 text-gray-900 dark:text-white hover:bg-gray-200 dark:hover:bg-white/10"
-          >
-            <X className="w-4 h-4" />
-          </Button>
-        </div>
-        
-        {/* Logo */}
-        <div className="p-6 border-b border-gray-200 dark:border-white/10">
-          <div className="flex items-center gap-3">
-            <div className="w-10 h-10 bg-gradient-to-r from-[#0db2f3] to-blue-500 rounded-xl flex items-center justify-center">
-              <Mic className="w-5 h-5 text-white" />
-            </div>
-            <div>
-              <h1 className="text-xl font-bold bg-gradient-to-r from-[#0db2f3] to-blue-400 bg-clip-text text-transparent">
-                ToriType
-              </h1>
-              <p className="text-xs text-gray-500 dark:text-white/60">Nigerian AI Platform</p>
-            </div>
-          </div>
-        </div>
-
-        {/* Menu Items */}
-        <nav className="flex-1 p-4">
-          <div className="space-y-2">
-            {menuItems.map((item) => {
-              const Icon = item.icon;
-              return (
-                <button
-                  key={item.id}
-                  className="w-full flex items-center gap-3 px-4 py-3 rounded-xl transition-all duration-300 text-gray-600 dark:text-white/70 hover:text-gray-900 dark:hover:text-white hover:bg-gray-100 dark:hover:bg-white/5"
-                >
-                  <Icon className="w-5 h-5" />
-                  <span className="font-medium">{item.label}</span>
-                </button>
-              );
-            })}
-          </div>
-        </nav>
-
-        {/* User Profile & Sign Out */}
-        <div className="p-4 border-t border-gray-200 dark:border-white/10">
-          <div className="flex items-center gap-3 mb-4 p-3 bg-gray-100 dark:bg-white/5 rounded-xl">
-            <Image
-              src={user?.user_metadata?.avatar_url || '/default-avatar.png'}
-              alt="Profile"
-              width={40}
-              height={40}
-              className="w-10 h-10 rounded-full border-2 border-[#0db2f3]/30"
-            />
-            <div className="flex-1 min-w-0">
-              <p className="text-sm font-medium text-gray-900 dark:text-white truncate">
-                {user?.user_metadata?.full_name || user?.email}
-              </p>
-              <p className="text-xs text-gray-600 dark:text-white/60 truncate">{user?.email}</p>
-            </div>
-          </div>
-          
-          {/* Theme Toggle */}
-          {mounted && (
-            <div className="mb-4">
-              <Button
-                onClick={() => setTheme(resolvedTheme === 'dark' ? 'light' : 'dark')}
-                variant="outline"
-                className="w-full bg-gray-100 dark:bg-white/5 border-gray-300 dark:border-white/10 text-gray-900 dark:text-white hover:bg-gray-200 dark:hover:bg-white/10 hover:border-gray-400 dark:hover:border-white/20"
-              >
-                {resolvedTheme === 'dark' ? (
-                  <>
-                    <Sun className="w-4 h-4 mr-2" />
-                    Light Mode
-                  </>
-                ) : (
-                  <>
-                    <Moon className="w-4 h-4 mr-2" />
-                    Dark Mode
-                  </>
-                )}
-              </Button>
-            </div>
-          )}
-          
-          <Button
-            onClick={handleSignOut}
-            variant="outline"
-            className="w-full bg-red-500/10 border-red-500/20 text-red-400 hover:bg-red-500/20 hover:border-red-500/30"
-          >
-            <LogOut className="w-4 h-4 mr-2" />
-            Sign Out
-          </Button>
-        </div>
-      </div>
+      {/* Sidebar */}
+      <Sidebar
+        user={user}
+        sidebarOpen={sidebarOpen}
+        setSidebarOpen={setSidebarOpen}
+        mounted={mounted}
+        handleSignOut={handleSignOut}
+      />
 
       {/* Main Content */}
       <div className="flex-1 flex flex-col lg:ml-0">
-        {/* Content Area */}
         <main className="flex-1 p-4 lg:p-8 pt-16 lg:pt-8 flex flex-col">
-          <div className="max-w-6xl mx-auto h-full flex flex-col">
+          <div className="max-w-6xl mx-auto h-full flex flex-col gap-6">
             
-            {/* Transcription Display - Top with more space */}
-            <div className="flex-1 min-h-0 bg-white/80 dark:bg-white/5 backdrop-blur-sm border border-gray-200 dark:border-white/10 rounded-2xl p-4 sm:p-6 mb-4 sm:mb-6">
-              <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between mb-4 gap-3">
-                <div className="flex items-center gap-3">
-                  <h2 className="text-lg sm:text-xl font-semibold text-gray-900 dark:text-white">Transcription</h2>
-                  
-                  {/* API Status Indicator */}
-                  <div className="flex items-center gap-2">
-                    <div className={`w-2 h-2 rounded-full ${
-                      apiStatus === 'connected' ? 'bg-green-500' :
-                      apiStatus === 'connecting' ? 'bg-yellow-500 animate-pulse' :
-                      apiStatus === 'error' ? 'bg-red-500' : 'bg-gray-400'
-                    }`}></div>
-                    <span className="text-xs text-gray-500 dark:text-gray-400">
-                      {apiStatus === 'connected' ? 'API Connected' :
-                       apiStatus === 'connecting' ? 'Connecting...' :
-                       apiStatus === 'error' ? 'Mock Mode' : 'Checking...'}
-                    </span>
-                  </div>
-                </div>
-                <div className="flex flex-wrap items-center gap-2">
-                  <Button
-                    onClick={handleCopyText}
-                    variant="outline"
-                    size="sm"
-                    className="bg-gray-100 dark:bg-white/10 border-gray-300 dark:border-white/20 text-gray-900 dark:text-white hover:bg-gray-200 dark:hover:bg-white/20"
-                  >
-                    {isCopied ? <Check className="w-4 h-4 mr-2" /> : <Copy className="w-4 h-4 mr-2" />}
-                    {isCopied ? 'Copied!' : 'Copy'}
-                  </Button>
-                  
-                  <Button
-                    onClick={downloadTranscript}
-                    variant="outline"
-                    size="sm"
-                    className="bg-gray-100 dark:bg-white/10 border-gray-300 dark:border-white/20 text-gray-900 dark:text-white hover:bg-gray-200 dark:hover:bg-white/20"
-                  >
-                    <Download className="w-4 h-4 mr-2" />
-                    Download
-                  </Button>
-                  
-                  {audioBlob && (
-                    <Button
-                      onClick={downloadAudio}
-                      variant="outline"
-                      size="sm"
-                      className="bg-gray-100 dark:bg-white/10 border-gray-300 dark:border-white/20 text-gray-900 dark:text-white hover:bg-gray-200 dark:hover:bg-white/20"
-                    >
-                      <FileAudio className="w-4 h-4 mr-2" />
-                      Audio
-                    </Button>
-                  )}
-                  
-                  {!isEditing ? (
-                    <Button
-                      onClick={handleEditStart}
-                      variant="outline"
-                      size="sm"
-                      className="bg-gray-100 dark:bg-white/10 border-gray-300 dark:border-white/20 text-gray-900 dark:text-white hover:bg-gray-200 dark:hover:bg-white/20"
-                    >
-                      <Edit3 className="w-4 h-4 mr-2" />
-                      Edit
-                    </Button>
-                  ) : (
-                    <div className="flex gap-2">
-                      <Button
-                        onClick={handleEditSave}
-                        size="sm"
-                        className="bg-green-600 hover:bg-green-700 text-white"
-                      >
-                        <Save className="w-4 h-4 mr-2" />
-                        Save
-                      </Button>
-                      <Button
-                        onClick={handleEditCancel}
-                        variant="outline"
-                        size="sm"
-                        className="bg-gray-100 dark:bg-white/10 border-gray-300 dark:border-white/20 text-gray-900 dark:text-white hover:bg-gray-200 dark:hover:bg-white/20"
-                      >
-                        <RotateCcw className="w-4 h-4 mr-2" />
-                        Cancel
-                      </Button>
-                    </div>
-                  )}
-                </div>
-              </div>
-              
-              {isEditing ? (
-                <textarea
-                  ref={textareaRef}
-                  value={editableText}
-                  onChange={(e) => setEditableText(e.target.value)}
-                  className="w-full h-full min-h-[300px] sm:min-h-[400px] p-4 bg-white dark:bg-gray-800 border border-gray-300 dark:border-gray-600 rounded-lg text-gray-900 dark:text-white resize-none focus:ring-2 focus:ring-[#0db2f3] focus:border-[#0db2f3] outline-none"
-                  placeholder="Your transcription will appear here..."
-                />
-              ) : (
-                <div className="h-full min-h-[300px] sm:min-h-[400px] p-4 bg-gray-50 dark:bg-gray-800/50 rounded-lg overflow-y-auto relative">
-                  {isProcessing && (
-                    <div className="absolute inset-0 bg-white/80 dark:bg-gray-800/80 flex items-center justify-center backdrop-blur-sm rounded-lg">
-                      <div className="flex items-center gap-3">
-                        <div className="animate-spin rounded-full h-6 w-6 border-b-2 border-[#0db2f3]"></div>
-                        <span className="text-gray-900 dark:text-white font-medium">Processing audio...</span>
-                      </div>
-                    </div>
-                  )}
-                  <p className="text-gray-900 dark:text-white whitespace-pre-wrap leading-relaxed text-sm sm:text-base">
-                    {transcriptText}
-                  </p>
-                </div>
-              )}
-            </div>
+            {/* Transcription Display */}
+            <TranscriptionDisplay
+              transcriptText={transcriptText}
+              isEditing={isEditing}
+              editableText={editableText}
+              setEditableText={setEditableText}
+              apiStatus={apiStatus}
+              isCopied={isCopied}
+              isProcessing={isProcessing}
+              audioBlob={audioBlob}
+              textareaRef={textareaRef}
+              handleCopyText={handleCopyText}
+              downloadTranscript={downloadTranscript}
+              downloadAudio={downloadAudio}
+              handleEditStart={handleEditStart}
+              handleEditSave={handleEditSave}
+              handleEditCancel={handleEditCancel}
+            />
 
-            {/* Control Buttons & Upload - Bottom */}
-            <div className="bg-white/80 dark:bg-white/5 backdrop-blur-sm border border-gray-200 dark:border-white/10 rounded-2xl p-4 sm:p-6">
-              
-              {/* Mode Toggle */}
-              <div className="flex flex-col lg:flex-row lg:items-center gap-6">
-                <div className="flex items-center justify-center">
-                  <div className="flex items-center gap-2 bg-gray-100 dark:bg-gray-800 rounded-xl p-1">
-                    <button
-                      onClick={() => setActiveMode('record')}
-                      className={`flex items-center gap-2 px-3 sm:px-4 py-2 rounded-lg transition-all text-sm sm:text-base ${
-                        activeMode === 'record'
-                          ? 'bg-[#0db2f3] text-white shadow-lg'
-                          : 'text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-white'
-                      }`}
-                    >
-                      <Mic className="w-4 h-4" />
-                      Record
-                    </button>
-                    <button
-                      onClick={() => setActiveMode('chat')}
-                      className={`flex items-center gap-2 px-3 sm:px-4 py-2 rounded-lg transition-all text-sm sm:text-base ${
-                        activeMode === 'chat'
-                          ? 'bg-[#0db2f3] text-white shadow-lg'
-                          : 'text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-white'
-                      }`}
-                    >
-                      <MessageCircle className="w-4 h-4" />
-                      Chat
-                    </button>
-                    <button
-                      onClick={() => setActiveMode('upload')}
-                      className={`flex items-center gap-2 px-3 sm:px-4 py-2 rounded-lg transition-all text-sm sm:text-base ${
-                        activeMode === 'upload'
-                          ? 'bg-[#0db2f3] text-white shadow-lg'
-                          : 'text-gray-600 dark:text-gray-400 hover:text-gray-900 dark:hover:text-white'
-                      }`}
-                    >
-                      <Upload className="w-4 h-4" />
-                      Upload
-                    </button>
-                  </div>
-                </div>
-
-                {/* Recording Controls */}
-                {activeMode === 'record' && (
-                  <div className="flex flex-col sm:flex-row items-center justify-center gap-4">
-                    {/* Timer */}
-                    <div className="text-xl sm:text-2xl font-mono font-bold text-gray-900 dark:text-white">
-                      {formatTime(recordingTime)}
-                    </div>
-
-                    {/* Recording Button */}
-                    {!isRecording ? (
-                      <Button
-                        onClick={startRecording}
-                        className="w-14 h-14 sm:w-16 sm:h-16 rounded-full bg-gradient-to-r from-[#0db2f3] to-blue-500 hover:from-[#0db2f3]/80 hover:to-blue-500/80 shadow-lg shadow-[#0db2f3]/30 transition-all duration-300 hover:scale-105"
-                      >
-                        <Mic className="w-5 h-5 sm:w-6 sm:h-6 text-white" />
-                      </Button>
-                    ) : (
-                      <div className="flex items-center gap-3">
-                        <Button
-                          onClick={pauseRecording}
-                          variant="outline"
-                          className="w-10 h-10 sm:w-12 sm:h-12 rounded-full border-gray-300 dark:border-white/20 text-gray-900 dark:text-white hover:bg-gray-100 dark:hover:bg-white/10"
-                        >
-                          {isPaused ? <Play className="w-4 h-4 sm:w-5 sm:h-5" /> : <Pause className="w-4 h-4 sm:w-5 sm:h-5" />}
-                        </Button>
-                        
-                        <Button
-                          onClick={stopRecording}
-                          className="w-12 h-12 sm:w-14 sm:h-14 rounded-full bg-red-500 hover:bg-red-600 shadow-lg shadow-red-500/30 transition-all duration-300"
-                        >
-                          <Square className="w-4 h-4 sm:w-5 sm:h-5 text-white fill-current" />
-                        </Button>
-                      </div>
-                    )}
-
-                    {/* Status */}
-                    {isRecording && (
-                      <div className="flex items-center gap-2 text-xs sm:text-sm text-gray-600 dark:text-gray-400">
-                        <div className="w-2 h-2 bg-red-500 rounded-full animate-pulse"></div>
-                        {isPaused ? 'Recording Paused' : 'Recording in Progress'}
-                      </div>
-                    )}
-                  </div>
-                )}
-
-                {/* Chat Interface */}
-                {activeMode === 'chat' && (
-                  <div className="flex items-center justify-center">
-                    <div className="text-gray-600 dark:text-gray-400 text-center">
-                      Chat functionality coming soon...
-                    </div>
-                  </div>
-                )}
-
-                {/* Upload Interface */}
-                {activeMode === 'upload' && (
-                  <div className="flex-1">
-                    <div 
-                      className={`border-2 border-dashed rounded-xl p-4 sm:p-6 text-center transition-colors cursor-pointer ${
-                        isDragOver 
-                          ? 'border-[#0db2f3] bg-[#0db2f3]/10 dark:bg-[#0db2f3]/20' 
-                          : 'border-gray-300 dark:border-gray-600 hover:border-[#0db2f3] dark:hover:border-[#0db2f3]'
-                      } ${isProcessing ? 'opacity-50 cursor-not-allowed' : ''}`}
-                      onDragOver={handleDragOver}
-                      onDragLeave={handleDragLeave}
-                      onDrop={handleDrop}
-                      onClick={triggerFileUpload}
-                    >
-                      <FileAudio className={`w-8 h-8 sm:w-10 sm:h-10 mx-auto mb-3 ${
-                        isDragOver ? 'text-[#0db2f3]' : 'text-gray-400 dark:text-gray-500'
-                      }`} />
-                      <p className="text-gray-600 dark:text-gray-400 mb-2 text-sm sm:text-base">
-                        {isDragOver ? 'Drop your audio file here!' : 'Drop your audio file here or click to browse'}
-                      </p>
-                      <p className="text-xs sm:text-sm text-gray-500 dark:text-gray-500 mb-4">Supports MP3, WAV, M4A, WEBM, OGG (Max 100MB)</p>
-                      
-                      {uploadedFileName && (
-                        <div className="mb-4 p-2 bg-green-100 dark:bg-green-900/30 rounded-lg">
-                          <p className="text-green-800 dark:text-green-400 text-sm">
-                            ✅ Last uploaded: {uploadedFileName}
-                          </p>
-                        </div>
-                      )}
-                      
-                      <input
-                        type="file"
-                        accept="audio/*"
-                        onChange={handleFileUpload}
-                        className="hidden"
-                        id="file-upload"
-                        disabled={isProcessing}
-                      />
-                      <label 
-                        htmlFor="file-upload"
-                        onClick={triggerFileUpload}
-                        className={`inline-flex items-center justify-center px-4 py-2 rounded-lg font-medium text-sm sm:text-base transition-all cursor-pointer ${
-                          isProcessing 
-                            ? 'bg-gray-400 text-white opacity-50 cursor-not-allowed'
-                            : 'bg-gradient-to-r from-[#0db2f3] to-blue-500 hover:from-[#0db2f3]/80 hover:to-blue-500/80 text-white shadow-lg hover:shadow-xl'
-                        }`}
-                      >
-                        {isProcessing ? (
-                          <>
-                            <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-white mr-2"></div>
-                            Processing...
-                          </>
-                        ) : (
-                          <>
-                            <Plus className="w-4 h-4 mr-2" />
-                            Choose File
-                          </>
-                        )}
-                      </label>
-                    </div>
-                  </div>
-                )}
-              </div>
-            </div>
+            {/* Control Panel */}
+            <ControlPanel
+              activeMode={activeMode}
+              setActiveMode={setActiveMode}
+              isRecording={isRecording}
+              isPaused={isPaused}
+              recordingTime={recordingTime}
+              isDragOver={isDragOver}
+              isProcessing={isProcessing}
+              uploadedFileName={uploadedFileName}
+              startRecording={startRecording}
+              pauseRecording={pauseRecording}
+              stopRecording={stopRecording}
+              handleDragOver={handleDragOver}
+              handleDragLeave={handleDragLeave}
+              handleDrop={handleDrop}
+              handleFileUpload={handleFileUpload}
+              triggerFileUpload={triggerFileUpload}
+            />
 
           </div>
         </main>
